requirements/base.txt
# Core Django
Django==5.0.1
django-environ==0.11.2
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Redis & Celery
redis==5.0.1
celery==5.3.4
django-celery-beat==2.7.0
django-celery-results==2.6.0
django-redis==5.4.0

# WebSockets
channels==4.0.0
channels-redis==4.1.0

# Security
django-ratelimit==4.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
django-modeltranslation==0.18.12

